refactor(settings): organize settings functions
