+++ src/renderer/src/features/SettingsHub/SettingsHub.tsx
@@ -1,65 +1,8 @@
-import log from "electron-log/renderer";
 import { Button, Stack } from "~/components";
-import * as dbUtilHooks from "~/hooks/ipc/useDatabaseUtilities";
-import * as dialogHooks from "~/hooks/ipc/useFileDialogs";
-import * as loggerHooks from "~/hooks/ipc/useLogger";
-import * as settingsHooks from "~/hooks/ipc/useSettingsUtilities";
-import { useToasts } from "../Toasts/useToasts";
+import { useSettingsFns } from "./hooks/useSettingsFns";
 
 export function SettingsHub() {
-  const loadAthletes = dialogHooks.useLoadAthletesFile();
-  const loadStation = dialogHooks.useLoadStationsFile();
-  const loadDNS = dialogHooks.useLoadDNSFile();
-  const loadDNF = dialogHooks.useLoadDNFFile();
-  const initializeDatabaseMutation = dbUtilHooks.useInitializeDatabase();
-  const clearDatabaseMutation = dbUtilHooks.useClearDatabase();
-  const importRunnersFile = dialogHooks.useImportRunnersFromCSV();
-  const resetAppSettingsMutation = settingsHooks.useResetAppSettings();
-  const rfidInitMutation = settingsHooks.useRFIDInitialize();
-  const { createToast } = useToasts();
-
-  const resetAppSettings = () => {
-    log.info("testing renderer to main log");
-    loggerHooks.useMainLogger("warning", "User click: Reset App Settings");
-    createToast({ message: "App Settings: Resetting", type: "info" });
-    resetAppSettingsMutation.mutate("resetAppSettings");
-  };
-
-  const rfidInitialize = () => {
-    createToast({ message: "App Settings: Resetting", type: "info" });
-    rfidInitMutation.mutate("resetAppSettings");
-  };
-
-  const importAthletesFile = () => {
-    createToast({ message: "Loading Athletes file", type: "info" });
-    loadAthletes.mutate("ping from the renderer!");
-  };
-
-  const importStationsFile = () => {
-    createToast({ message: "Loading Stations file", type: "info" });
-    loadStation.mutate("ping from the renderer!");
-  };
-
-  const importDNSFile = () => {
-    createToast({ message: "Loading DNS file", type: "info" });
-    loadDNS.mutate("ping from the renderer!");
-  };
-
-  const importDNFFile = () => {
-    createToast({ message: "Loading DNF file", type: "info" });
-    loadDNF.mutate("ping from the renderer!");
-  };
-
-  const importRunnerCSVFile = () => {
-    createToast({ message: "Importing from CSV file", type: "info" });
-    importRunnersFile.mutate("ping from the renderer!");
-  };
-
-  const initializeDatabase = () => {
-    createToast({ message: "Reinitializing database!", type: "danger" });
-    clearDatabaseMutation.mutate("Clearing Database of all tables!");
-    initializeDatabaseMutation.mutate("Initializing Database!");
-  };
+  const settingsFns = useSettingsFns();
 
   return (
     <div>
@@ -70,16 +13,16 @@ export function SettingsHub() {
         <Stack direction="row" align="stretch">
           <Stack direction="col">
             <b>Station Setup</b>
-            <Button color="primary" size="md" onClick={importStationsFile}>
+            <Button color="primary" size="md" onClick={settingsFns.importStationsFile}>
               Load Stations File
             </Button>
-            <Button color="primary" size="md" onClick={importAthletesFile}>
+            <Button color="primary" size="md" onClick={settingsFns.importAthletesFile}>
               Load Athletes File
             </Button>
-            <Button color="primary" size="md" onClick={importDNSFile}>
+            <Button color="primary" size="md" onClick={settingsFns.importDNSFile}>
               Load DNS File
             </Button>
-            <Button color="primary" size="md" onClick={importDNFFile}>
+            <Button color="primary" size="md" onClick={settingsFns.importDNFFile}>
               Load DNF File
             </Button>
           </Stack>
@@ -89,13 +32,13 @@ export function SettingsHub() {
         <Stack direction="row" align="stretch">
           <Stack direction="col">
             <b>RFID Configuration [Start Line only]</b>
-            <Button color="neutral" variant="outlined" onClick={rfidInitialize}>
+            <Button color="neutral" variant="outlined" onClick={settingsFns.rfidInitialize}>
               Initialize RFID
             </Button>
           </Stack>
           <Stack direction="col">
             <b>App Settings</b>
-            <Button color="warning" variant="outlined" onClick={resetAppSettings}>
+            <Button color="warning" variant="outlined" onClick={settingsFns.resetAppSettings}>
               Reset App Settings
             </Button>
           </Stack>
@@ -104,10 +47,10 @@ export function SettingsHub() {
       <div>
         <Stack direction="col">
           <b>Developer Tools</b>
-          <Button color="danger" variant="outlined" onClick={initializeDatabase}>
+          <Button color="danger" variant="outlined" onClick={settingsFns.initializeDatabase}>
             Destroy & Reinitialize Database
           </Button>
-          <Button color="danger" variant="outlined" onClick={importRunnerCSVFile}>
+          <Button color="danger" variant="outlined" onClick={settingsFns.importRunnerCSVFile}>
             Recover Data from CSV File
           </Button>
         </Stack>
